+++ app/views/home/index.html.erb
@@ -12,7 +12,9 @@
       </p>
       <p class="description">
         <span><%= @weather[:description] %></span>
-        <%= image_tag @weather[:icon_url] %>
+        <% if @weather[:icon_url] %>
+          <%= image_tag @weather[:icon_url] %>
+        <% end %>
       </p>
     </div>
     <div class="weather-details">
